[Refact] : KickboardDetailViewController 리팩토링
